requirements.txt
pygame>=2.5.0
pyyaml>=6.0
requests>=2.31.0
